Add files via upload
I have members for 8 groups chosen. Tried to build a graph in 2 variants: 1. group is a node, edges - intersections of members (obviously as few members - no intersections), so I looked into the 2nd variant ->2. group is a node, edges - intersections of cities (so let's say people in Uhta likes a lot 2 groups - 'humor' and 'strange sth')

I was able to draw a plot for the 1st case, but not for the 2nd, although the code is the same. The underline data in members_result is different